+++ CHANGELOG.md
@@ -12,7 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
-* Nothing
+* fix: ```generate-circleci-config.py``` had oustanding problems from Python 2.7 -> 3.7
 
 ### Removed